requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
# To install the required packages, run the following command:
#
#    pip install -r requirements.txt
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.47.2
    # via matplotlib
imagecodecs==2024.1.1
    # via -r requirements.in
imageio==2.33.1
    # via scikit-image
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
matplotlib==3.8.2
    # via
    #   -r requirements.in
    #   seaborn
networkx==3.2.1
    # via scikit-image
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   imagecodecs
    #   imageio
    #   matplotlib
    #   pandas
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tifffile
packaging==23.2
    # via
    #   matplotlib
    #   plotly
    #   scikit-image
pandas==2.2.0
    # via
    #   -r requirements.in
    #   seaborn
pillow==10.2.0
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   scikit-image
plotly==5.18.0
    # via -r requirements.in
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
scikit-image==0.22.0
    # via -r requirements.in
scikit-learn==1.4.0
    # via -r requirements.in
scipy==1.12.0
    # via
    #   scikit-image
    #   scikit-learn
seaborn==0.13.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tenacity==8.2.3
    # via plotly
threadpoolctl==3.2.0
    # via scikit-learn
tifffile==2024.1.30
    # via scikit-image
tzdata==2023.4
    # via pandas
